requirements.txt
paramiko>=3.3.1,<4
pyVmomi
pyVim
lxml
certvalidator
asn1crypto
tlslite-ng>=0.8.0a45,<0.8.3
ecdsa
